refactor(credit): modularize sponsor_engine → sponsor/ package (Wave 1 PR #2)
Converts sponsor_engine.py (390 LOC) into a 2-file package:
  sponsor/
    person_extraction.py — extract_key_persons_from_analysis (deterministic)
    service.py           — analyze_sponsor (orchestrator, never-raises)
    __init__.py          — standard imports, __all__

No models.py — keeps existing dict returns (no premature dataclass formalization).
Never-raises contract: service.py wraps all exceptions, returns NOT_ASSESSED.
External deps (prompt_registry, evidence_selector) moved to lazy imports in service.
stdlib logging → structlog.

Callers updated: deep_review.py (2 sites), kyc_pipeline_screening.py (1 site).
test_vertical_engines.py EXPECTED_MODULES updated.
Old flat file deleted.
