+++ docs/plans/2026-03-15-refactor-credit-vertical-modular-alignment-wave1-plan.md
@@ -154,7 +154,7 @@ critic/
 
 ---
 
-### Phase 2: sponsor/ (PR #2)
+### Phase 2: sponsor/ (PR #2) ✅
 
 **Source:** `sponsor_engine.py` (390 LOC)
